# schema/openapi/openapi.yaml
openapi: '3.0.3'
info:
  title: Backbone API
  version: '1.0.0'
  description: |
    Auto-generated by backbone-cli.
    This API provides 11 standard CRUD endpoints for each model.
servers:
  - url: http://localhost:3000
    description: Development server
...
paths:
...
  /api/v1/category_costing_defaultses:
    get:
      tags:
        - CategoryCostingDefaultses
      summary: List CategoryCostingDefaultses
      description: Get a paginated list of CategoryCostingDefaultses with optional filtering and sorting
      operationId: listCategoryCostingDefaultses
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoryCostingDefaultsList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - CategoryCostingDefaultses
      summary: Create CategoryCostingDefaults
      description: Create a new CategoryCostingDefaults
      operationId: createCategoryCostingDefaults
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCategoryCostingDefaultsRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoryCostingDefaults'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/v1/repair_orders:
    get:
      tags:
        - RepairOrders
      summary: List RepairOrders
      description: Get a paginated list of RepairOrders with optional filtering and sorting
      operationId: listRepairOrders
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RepairOrderList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - RepairOrders
      summary: Create RepairOrder
      description: Create a new RepairOrder
      operationId: createRepairOrder
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRepairOrderRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RepairOrder'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  schemas:
...
    CategoryCostingDefaults:
      type: object
      required:
        - id
        - company_id
        - product_category_id
        - metadata
      properties:
        id:
          type: string
          format: uuid
        company_id:
          type: string
          format: uuid
        product_category_id:
          type: string
          format: uuid
        wip_account_id:
          nullable: true
          type: string
          format: uuid
        fg_account_id:
          nullable: true
          type: string
          format: uuid
        raw_material_account_id:
          nullable: true
          type: string
          format: uuid
        conversion_cost_account_id:
          nullable: true
          type: string
          format: uuid
        subcontract_interim_account_id:
          nullable: true
          type: string
          format: uuid
        cost_variance_account_id:
          nullable: true
          type: string
          format: uuid
        inventory_loss_account_id:
          nullable: true
          type: string
          format: uuid
        repair_expense_account_id:
          nullable: true
          type: string
          format: uuid
        metadata:
          type: object
...
    CreateCategoryCostingDefaultsRequest:
      type: object
      required:
        - company_id
        - product_category_id
        - metadata
      properties:
        company_id:
          type: string
          format: uuid
        product_category_id:
          type: string
          format: uuid
        wip_account_id:
          nullable: true
          type: string
          format: uuid
        fg_account_id:
          nullable: true
          type: string
          format: uuid
        raw_material_account_id:
          nullable: true
          type: string
          format: uuid
        conversion_cost_account_id:
          nullable: true
          type: string
          format: uuid
        subcontract_interim_account_id:
          nullable: true
          type: string
          format: uuid
        cost_variance_account_id:
          nullable: true
          type: string
          format: uuid
        inventory_loss_account_id:
          nullable: true
          type: string
          format: uuid
        repair_expense_account_id:
          nullable: true
          type: string
          format: uuid
        metadata:
          type: object
...
    CategoryCostingDefaultsList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/CategoryCostingDefaults'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
    RepairOrder:
      type: object
      required:
        - id
        - company_id
        - repair_number
        - item_id
        - quantity
        - status
        - metadata
      properties:
        id:
          type: string
          format: uuid
        company_id:
          type: string
          format: uuid
        repair_number:
          type: string
          maxLength: 40
        item_id:
          type: string
          format: uuid
        product_category_id:
          nullable: true
          type: string
          format: uuid
        quantity:
          type: string
        status:
          type: string
        metadata:
          type: object
...
    CreateRepairOrderRequest:
      type: object
      required:
        - company_id
        - repair_number
        - item_id
        - quantity
        - status
        - metadata
      properties:
        company_id:
          type: string
          format: uuid
        repair_number:
          type: string
          maxLength: 40
        item_id:
          type: string
          format: uuid
        product_category_id:
          nullable: true
          type: string
          format: uuid
        quantity:
          type: string
        status:
          type: string
        metadata:
          type: object
...
    RepairOrderList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/RepairOrder'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
...
    PaginationMeta:
      type: object
      properties:
        total:
          type: integer
          example: 100
        page:
          type: integer
          example: 1
        per_page:
          type: integer
          example: 20
        total_pages:
          type: integer
          example: 5
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
          example: Invalid input data
        details:
          type: object
          additionalProperties: true
  parameters:
...
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PerPageParam:
      name: per_page
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    SortByParam:
      name: sort_by
      in: query
      schema:
        type: string
    SortOrderParam:
      name: sort_order
      in: query
      schema:
        type: string
        enum: [asc, desc]
        default: asc
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []